修复:更新GeneralOpenAIRequest字段以使用指针类型并调整转换逻辑

+++ relay/channel/xunfei/relay-xunfei.go
@@ -49,7 +49,7 @@ func requestOpenAI2Xunfei(request dto.GeneralOpenAIRequest, xunfeiAppId string, 
 	xunfeiRequest.Header.AppId = xunfeiAppId
 	xunfeiRequest.Parameter.Chat.Domain = domain
 	xunfeiRequest.Parameter.Chat.Temperature = request.Temperature
-	xunfeiRequest.Parameter.Chat.TopK = lo.FromPtrOr(request.N, 0)
+	xunfeiRequest.Parameter.Chat.TopK = int(lo.FromPtrOr(request.N, uint(0)))
 	xunfeiRequest.Parameter.Chat.MaxTokens = request.GetMaxTokens()
 	xunfeiRequest.Payload.Message.Text = messages
 	return &xunfeiRequest
